refactor: remove server type re-exports and side-effect import — boundary clean
- Remove backward-compat export type re-exports from campaigns.ts, notes.ts, gmscreens.ts
- Remove WINDOW_STATES/WindowState re-exports from GMScreen model (client backward compat only)
- Remove side-effect import of ~/server/functions/gmscreens from play.tsx (no longer needed)
- Update test files to import types directly from ~/types/ instead of server files
- Boundary verification: zero ~/server/ imports in hooks/components/utils/providers/services
