+++ app/server/db/models/GMScreen.ts
@@ -2,8 +2,6 @@ import mongoose from 'mongoose'
 import { WINDOW_STATES } from '~/types/gmscreen'
 import type { WindowState } from '~/types/gmscreen'
 
-export { WINDOW_STATES }
-export type { WindowState }
 
 // ---------------------------------------------------------------------------
 // Constants – practical guardrails for embedded arrays
