+++ tests/server/functions/gmscreens.test.ts
@@ -95,7 +95,7 @@ import {
   removeStackItemSchema,
   SUPPORTED_COLLECTIONS,
 } from '~/server/functions/gmscreens'
-import type { GMScreenData, GMScreenDetailData, WindowData, StackData, StackItemData } from '~/server/functions/gmscreens'
+import type { GMScreenData, GMScreenDetailData, WindowData, StackData, StackItemData } from '~/types/gmscreen'
 import { removeDocumentRefsFromScreens } from '~/server/functions/gmscreens-helpers'
 import { serverCaptureEvent, serverCaptureException } from '~/server/utils/posthog'
